% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Rapier2DWorld}
\alias{Rapier2DWorld}
\alias{$.Rapier2DWorld}
\title{The 2D World of Rapier}
\format{
An object of class \code{environment} of length 6.
}
\usage{
Rapier2DWorld
}
\description{
The 2D World of Rapier
}
\keyword{datasets}
